Regroup results component
