+++ src/components/results/angleResults/daily/dailySuccessTilt.js
@@ -3,10 +3,10 @@ import axios from 'axios';
 import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
 import { Chart } from 'primereact/components/chart/Chart';
-import CustomCard from '../../shared/card';
-import { URL } from '../../../redux/applicationReducer';
-import { T } from '../../../utilities/translator';
-import '../../../styles/results.css';
+import CustomCard from '../../../shared/card';
+import { URL } from '../../../../redux/applicationReducer';
+import { T } from '../../../../utilities/translator';
+import '../../../../styles/results.css';
 
 class DailySuccessTilt extends Component {
   static propTypes = {
