+++ src/components/results/angleResults/monthly/monthlyAngleDistribution.js
@@ -9,10 +9,10 @@ import { connect } from 'react-redux';
 import { Chart } from 'primereact/components/chart/Chart';
 import PropTypes from 'prop-types';
 import axios from 'axios';
-import CustomCard from '../../shared/card';
-import { T } from '../../../utilities/translator';
-import { URL } from '../../../redux/applicationReducer';
-import '../../../styles/results.css';
+import CustomCard from '../../../shared/card';
+import { T } from '../../../../utilities/translator';
+import { URL } from '../../../../redux/applicationReducer';
+import '../../../../styles/results.css';
 
 class MonthlyAngleDistribution extends Component {
   static propTypes = {
