+++ src/components/results/pressureResults/daily/recGoalProgress.js
@@ -9,8 +9,8 @@ import React, { Component } from 'react';
 import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
 import { ProgressBar } from 'primereact/components/progressbar/ProgressBar';
-import { T } from '../../../utilities/translator';
-import CustomCard from '../../shared/card';
+import { T } from '../../../../utilities/translator';
+import CustomCard from '../../../shared/card';
 
 class RecGoalProgress extends Component {
   static propTypes = {
