+++ src/components/results/pressureResults/monthly/monthlySittingTime.js
@@ -9,9 +9,9 @@ import { connect } from 'react-redux';
 import { Chart } from 'primereact/components/chart/Chart';
 import PropTypes from 'prop-types';
 import axios from 'axios';
-import CustomCard from '../../shared/card';
-import { T } from '../../../utilities/translator';
-import { URL } from '../../../redux/applicationReducer';
+import CustomCard from '../../../shared/card';
+import { T } from '../../../../utilities/translator';
+import { URL } from '../../../../redux/applicationReducer';
 
 class MonthlySittingTime extends Component {
   static propTypes = {
