+++ src/views/angleResults.js
@@ -10,13 +10,10 @@ import React, { Component } from 'react';
 
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
-import DailyAngleDistribution from '../components/results/angleResults/dailyAngleDistribution';
-import DailySuccessTilt from '../components/results/angleResults/dailySuccessTilt';
-import MonthlyAngleDistribution from '../components/results/angleResults/monthlyAngleDistribution';
-import MonthlySuccessTilt from '../components/results/angleResults/monthlySuccessTilt';
 import ResultsCalendar from '../components/results/resultsCalendar';
 import { T } from '../utilities/translator';
-import { IS_TABLET } from '../redux/applicationReducer';
+import MonthlyAngleResults from '../components/results/angleResults/monthly/monthlyAngleResults';
+import DailyAngleResults from '../components/results/angleResults/daily/dailyAngleResults';
 
 
 class AngleResults extends Component {
@@ -58,56 +55,14 @@ class AngleResults extends Component {
         <ResultsCalendar onPeriodChange={this.changePeriod} onDateChange={this.changeDate} onMonthChange={this.changeMonth} />
         <h2 className="center">{T.translate(`results.categories.angle.${this.props.language}`)}</h2>
         <hr />
-        {!IS_TABLET
-          && (
-          <div className="col-lg-2 leftMenu">
-            {this.state.period === 'day'
-              ? (
-                <ul className="graphlist">
-                  <li className="graphLink"><a href="results/angle#dailyAngle">{T.translate(`results.graphicsLink.angle.${this.props.language}`)}</a></li>
-                  <li className="graphLink"><a href="results/angle#dailyTilt">{T.translate(`SuccessfulTilt.tiltMade.${this.props.language}`)}</a></li>
-                </ul>
-              )
-              : (
-                <ul className="graphlist">
-                  <li className="graphLink"><a href="results/angle#monthlyAngle">{T.translate(`results.graphicsLink.angle.${this.props.language}`)}</a></li>
-                  <li className="graphLink"><a href="results/angle#monthlyTilt">{T.translate(`SuccessfulTilt.tiltMade.${this.props.language}`)}</a></li>
-                </ul>
-              )
-            }
-          </div>
+        {this.state.period === 'day'
+          ? (
+            <DailyAngleResults date={this.state.date} />
+          )
+          : (
+            <MonthlyAngleResults month={this.state.month} />
           )
         }
-        <div className=" col-lg-10 offset-lg-2 results resultsContainer">
-          <div className="col-lg-8 offset-lg-1">
-            {this.state.period === 'day'
-              ? (
-                <div>
-                  {this.state.date
-                  && (
-                  <div>
-                    <DailyAngleDistribution date={this.state.date} />
-                    <DailySuccessTilt date={this.state.date} />
-                  </div>
-                  )
-                }
-                </div>
-              )
-              : (
-                <div>
-                  {this.state.month
-                  && (
-                  <div>
-                    <MonthlyAngleDistribution month={this.state.month} />
-                    <MonthlySuccessTilt month={this.state.month} />
-                  </div>
-                  )
-                }
-                </div>
-              )
-            }
-          </div>
-        </div>
       </div>
     );
   }
